requirements.txt
asgiref==3.2.7
Django==3.0.4
markdown2==2.3.8
Pillow==7.0.0
Pygments==2.6.1
pytz==2019.3
sqlparse==0.3.1
